+++ .vscode/settings.json
@@ -3,6 +3,9 @@
         "sdl_ttf.h": "c",
         "stdlib.h": "c",
         "stdint.h": "c",
-        "sdl_timer.h": "c"
+        "sdl_timer.h": "c",
+        "sound.h": "c",
+        "stdio.h": "c",
+        "sdl.h": "c"
     }
 }
