+++ inc/input.h
@@ -4,8 +4,6 @@
 #include <stdlib.h>
 #include <stdint.h>
 
-extern int32_t game_is_running;
-
 void process_input();
 void get_entity_for_input();
